refactor to (probably) work in the browser using UMD boilerplate
The idea is that if bleach.js is loaded in a browser, it will either
contribute 'Bleach' as a global if there is no AMD support, or
use define() if there is AMD support.  And if it's in node, it just
does the node thing, although we also rely on the existing node
index.js magic to let the node binding automatically plug jsdom in
and use some hacks to make that work.

mocha tests still pass with this, but I have not yet run anything in
the browser.